requirements.txt
accelerate==0.25.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.40.2
build==1.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
colored==2.2.4
coloredlogs==15.0.1
comm==0.2.2
cuda-python==12.5.0
datasets==2.18.0
debugpy==1.8.6
decorator==5.1.1
diffusers==0.27.0
dill==0.3.8
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.2.0
h5py==3.10.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.6
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
janus==1.0.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.3.3
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==8.9.4.25
nvidia-modelopt==0.15.1
onnx==1.16.2
onnxruntime==1.19.2
optimum==1.17.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft==0.4.0
pillow==10.4.0
platformdirs==4.3.6
polygraphy==0.49.9
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
PuLP==2.9.0
pure_eval==0.2.3
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycuda==2024.1.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pynvml==11.5.3
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
matplotlib
